packages/api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=2.0.0
asyncpg>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pydantic>=2.0.0
loguru>=0.6.0
minio>=7.1.0
numpy>=1.21.0
pandas>=1.3.0
pyedflib>=0.1.0
prometheus-client>=0.14.0
opentelemetry-api>=1.11.0
opentelemetry-sdk>=1.11.0
opentelemetry-exporter-otlp>=1.11.0
opentelemetry-instrumentation-fastapi>=0.30b1
strawberry-graphql[fastapi]>=0.205.0
python-dateutil>=2.8.2
typing-extensions>=4.5.0
